🚚 Update directory structure
