+++ src/config.ts
@@ -1,0 +1,14 @@
+export const colors = {
+  ivory: 'rgb(240, 227, 206)',
+  buff: 'rgb(179, 147, 105)',
+  dark: 'rgb(30, 30, 30)',
+  red: 'rgb(200, 0, 0)',
+  blue: 'rgb(0, 0, 200)',
+  grey: 'rgb(150, 150, 150)',
+  safe: 'rgb(121, 202, 68)',
+};
+
+export const scales = {
+  squareSize: 9 / 80,
+  margin: 1 / 20,
+};
